requirements-dev.txt
pytest-qgis~=1.0.1
pytest-timeout==1.4.2
